Update for 1.6.2 (mostly)
Adds HardenedClay, CoalBlock, HayBale, StainedClay, HorseArmors, NameTag.
Adds new Horse and variants
new Packets and Packet Changes, all health and damage is now sent as floats.
templated HorseInventory

Fixed ProtocolTests so there are no more missing message tests.
Fixed Beacon and CommandBlock Messages not having proper equals/toString methods.
Fixed QuartzPillar not being rotational, added new RotationalSolid for Quartz/Logs.

+++ src/main/java/org/spout/vanilla/component/entity/living/Human.java
@@ -257,10 +257,10 @@ public class Human extends Living {
 		return getOwner().getData().get(VanillaData.IS_FLYING);
 	}
 
-	public void setFlyingSpeed(byte speed, boolean updateClient) {
+	public void setFlyingSpeed(float speed, boolean updateClient) {
 		Number value = getOwner().getData().put(VanillaData.FLYING_SPEED, speed);
 
-		byte previous = value == null ? VanillaData.FLYING_SPEED.getDefaultValue().byteValue() : value.byteValue();
+		float previous = value == null ? VanillaData.FLYING_SPEED.getDefaultValue().byteValue() : value.byteValue();
 
 		if (callAbilityChangeEvent().isCancelled()) {
 			getOwner().getData().put(VanillaData.FLYING_SPEED, previous);
@@ -269,16 +269,16 @@ public class Human extends Living {
 		updateAbilities(updateClient);
 	}
 
-	public void setFlyingSpeed(byte speed) {
+	public void setFlyingSpeed(float speed) {
 		setFlyingSpeed(speed, true);
 	}
 
-	public byte getFlyingSpeed() {
-		return getOwner().getData().get(VanillaData.FLYING_SPEED).byteValue();
+	public float getFlyingSpeed() {
+		return getOwner().getData().get(VanillaData.FLYING_SPEED).floatValue();
 	}
 
-	public void setWalkingSpeed(byte speed, boolean updateClient) {
-		byte previous = getOwner().getData().put(VanillaData.WALKING_SPEED, speed).byteValue();
+	public void setWalkingSpeed(float speed, boolean updateClient) {
+		float previous = getOwner().getData().put(VanillaData.WALKING_SPEED, speed).floatValue();
 		if (callAbilityChangeEvent().isCancelled()) {
 			getOwner().getData().put(VanillaData.WALKING_SPEED, previous);
 			return;
@@ -286,12 +286,12 @@ public class Human extends Living {
 		updateAbilities(updateClient);
 	}
 
-	public void setWalkingSpeed(byte speed) {
+	public void setWalkingSpeed(float speed) {
 		setWalkingSpeed(speed, true);
 	}
 
-	public byte getWalkingSpeed() {
-		return getOwner().getData().get(VanillaData.WALKING_SPEED).byteValue();
+	public float getWalkingSpeed() {
+		return getOwner().getData().get(VanillaData.WALKING_SPEED).floatValue();
 	}
 
 	public void setCanFly(boolean canFly, boolean updateClient) {
