+++ src/main/java/org/spout/vanilla/event/entity/HumanAbilityChangeEvent.java
@@ -35,8 +35,8 @@ import org.spout.vanilla.data.GameMode;
 
 public class HumanAbilityChangeEvent extends AbstractEntityEvent implements Cancellable {
 	private static final HandlerList handlers = new HandlerList();
-	private final byte flyingSpeed;
-	private final byte walkingSpeed;
+	private final float flyingSpeed;
+	private final float walkingSpeed;
 	private final boolean godMode;
 	private final boolean isFlying;
 	private final boolean canFly;
@@ -52,11 +52,11 @@ public class HumanAbilityChangeEvent extends AbstractEntityEvent implements Canc
 		creativeMode = human.getGameMode() == GameMode.CREATIVE;
 	}
 
-	public byte getFlyingSpeed() {
+	public float getFlyingSpeed() {
 		return flyingSpeed;
 	}
 
-	public byte getWalkingSpeed() {
+	public float getWalkingSpeed() {
 		return walkingSpeed;
 	}
